// src/lib/usergroup-api/models/data_UserGroupListResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { data_UserGroupListItemVO } from './data_UserGroupListItemVO';
export type data_UserGroupListResponse = {
    items?: Array<data_UserGroupListItemVO>;
    page?: number;
    pageSize?: number;
    total?: number;
};