// src/lib/usergroup-api/models/data_UserGroupVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { data_RuleConfig } from './data_RuleConfig';
export type data_UserGroupVO = {
    createdAt?: string;
    id?: number;
    name?: string;
    ruleConfig?: data_RuleConfig;
    status?: string;
    updatedAt?: string;
};